requirements.txt
psycopg2-binary>=2.9.10
pandas>=2.2.0
numpy>=1.26.4
python-dotenv>=1.0.1
ijson>=3.2.0
ijson>=3.2.0
